moved all company hierarchy reltaed use cases into CompanyHierarchyService, also updated the tests to use this new service instead of user service, now need to manage how we deal with invites in assign and accept/reject
